requirements.txt
fastapi==0.75.1
sqlalchemy==1.4.35
pymysql==1.0.2
cryptography==36.0.2
uvicorn==0.17.6
dictalchemy==0.1.2.7
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.5
bcrypt==3.2.0
loguru==0.6.0